move rpc code to its own module
